chore: extracting the Operation class into its own file
